src/main/java/com/taskmanagement/demo/service/TaskService.java
package com.taskmanagement.demo.service;

import java.util.List;
import java.util.concurrent.CompletableFuture;

import org.springframework.beans.factory.annotation.Autowired;
import org.springframework.scheduling.annotation.Async;
import org.springframework.stereotype.Service;

import com.taskmanagement.demo.model.Task;
import com.taskmanagement.demo.repository.TaskRepository;

@Service
public class TaskService {
    @Autowired
    private TaskRepository taskRepository;

    public List<Task> getAllTasks() {
        return taskRepository.findAll();
    }

    public Task getTaskById(Long id) {
        return taskRepository.findById(id).orElse(null);
    }

    public Task createTask(Task task) {
        return taskRepository.save(task);
    }

    public Task updateTask(Long id, Task taskDetails) {
        Task task = taskRepository.findById(id).orElse(null);
        if (task != null) {
            task.setTitle(taskDetails.getTitle());
            task.setDescription(taskDetails.getDescription());
            task.setStatus(taskDetails.getStatus());
            task.setDueDate(taskDetails.getDueDate());
            return taskRepository.save(task);
        }
        return null;
    }

    public void deleteTask(Long id) {
        taskRepository.deleteById(id);
    }

    @Async
    public CompletableFuture<List<Task>> getAllTasksAsync() {
        return CompletableFuture.completedFuture(taskRepository.findAll());
    }

    @Async
    public CompletableFuture<Task> getTaskByIdAsync(Long id) {
        return CompletableFuture.completedFuture(taskRepository.findById(id).orElse(null));
    }
}
